requirements.txt
flask==3.0.2
requests==2.31.0
pycryptodome==3.20.0
py_ecc==7.0.1
eth_keys==0.6.0
eth_utils==4.1.1
numpy==1.26.4
pytest==8.0.2
